Remove all 'sup' tags being passed to DOMPurify
The 'ALLOWED_TAGS' is actualy an extensive list of what html tags will
be allowed. It overides the default list from DOMPurify meaning that
only the sup tag will be allowed to pass through.

This reverts those changes (as they break a big chunk of the markdown on
the site)

I have added the sup tag to bootstrap.js allowed tags, for where
DOMPurify is not being used